Third time around (and use the next branch for collaboration)
There is (obviously, from the recent commit flaws) a need to keep
a volatile branch for collaboration, to allow for digested commits
in master. The next branch will rebase occasionally, so one cannot
rely on commit ids from that branch.

This patch corrects the commit mistake for the third time, to never
change that line again.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>