Unify detection of finger configuration changes
This patch simplifies and unifies the means by which finger
configuration changes are detected. The logic is almost the same,
except for the unlikely case of very quickly switching the clicking
finger, after first having moved it above the clicking area. This case
should now properly be detected as a pointing finger.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>