Correctly report zero fingers
This bug has been around since before alpha1, and yet the topic
of commit c7ed4c, prior to alpha1, was about precisely this; With
the bug, the last tracking id will sometimes be continued with a
new touch, resulting in a cursor jump.

This patch correctly sets zero fingers in HWData after receiving
a zero-finger event.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>