Add resting thumb detection to MTFinger
Robust thumb detection is essential for more advanced gestures, thumbs
being responsible for many false positives. One major feature of the
thumb is that its shape is oval, regardless of applied pressure. An
index finger pressed hard against the surface also turns oval, but the
touch/width ratio then increases significantly. This suggests that it
is possible to uniquely identify a resting thumb as a finger with oval
shape and normal touch/with ratio.

This patch adds detection of resting thumbs to the MTFinger structure.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>