refactor: Introduce mtdev
With the addition of the kernel MT slots, the MT event protocol
is capable of efficiently propagating changes to a set of tracked
contacts. At the same time, the need to treat a variety of different
kernel drivers is increased. This patch introduces the mtdev, an
abstract MT device which converts all valid MT event formats into
a uniform, slotted type B event stream. A command-line test program
is included.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>